make the comparison unsigned
svn path=/branches/KDE_3_3_BRANCH/kdegraphics/kpdf/; revision=358456

+++ xpdf/Catalog.cc
@@ -65,8 +65,8 @@ Catalog::Catalog(XRef *xrefA) {
   }
   pagesSize = numPages0 = (int)obj.getNum();
   obj.free();
-  if ((pagesSize >= INT_MAX / (signed) sizeof(Page *)) ||
-      (pagesSize >= INT_MAX / (signed) sizeof(Ref)))
+  if (((unsigned) pagesSize >= INT_MAX / sizeof(Page *)) ||
+      ((unsigned) pagesSize >= INT_MAX / sizeof(Ref)))
   {
     error(-1, "Invalid 'pagesSize'");
     ok = gFalse;
@@ -200,8 +200,8 @@ int Catalog::readPageTree(Dict *pagesDict, PageAttrs *attrs, int start) {
       }
       if (start >= pagesSize) {
 	pagesSize += 32;
-        if (pagesSize >= INT_MAX / (signed) sizeof(Page*) ||
-            pagesSize >= INT_MAX / (signed) sizeof(Ref)) {
+        if ((unsigned) pagesSize >= INT_MAX / sizeof(Page*) ||
+            (unsigned) pagesSize >= INT_MAX / sizeof(Ref)) {
           error(-1, "Invalid 'pagesSize' parameter.");
           goto err3;
         }
